+++ irs-api/src/main/java/org/eclipse/tractusx/irs/edc/AssemblyPartRelationship.java
@@ -26,7 +26,6 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -55,7 +54,7 @@ class AssemblyPartRelationship extends RelationshipSubmodel {
     public List<Relationship> asRelationships() {
         return Optional.ofNullable(this.childParts).stream().flatMap(Collection::stream)
                        .map(childData -> childData.toRelationship(this.catenaXId))
-                       .collect(Collectors.toList());
+                       .toList();
     }
 
     /**